Fix compatibility with ARM
